refactor(profile): moves component to module
