+++ src/modules/Profile/Addresses.tsx
@@ -1,15 +1,15 @@
 import { Grid, Container } from '@material-ui/core'
 import React from 'react'
 
-import Order from '../../modules/Cards/Order'
+import Address from '../Cards/Address'
 
-const OrderHistory: React.FC = () => {
+const Addresses: React.FC = () => {
   return (
     <Container>
       <Grid container spacing={2}>
         {[0, 1, 2].map((value) => (
           <Grid item xs key={value}>
-            <Order />
+            <Address isDefault={value === 0} />
           </Grid>
         ))}
       </Grid>
@@ -17,4 +17,4 @@ const OrderHistory: React.FC = () => {
   )
 }
 
-export default OrderHistory
+export default Addresses
